Long funds -> BigDec funds

+++ src/main/java/ru/aston/testproj/domain/dto/account/AccountDepositDto.java
@@ -6,6 +6,8 @@ import static ru.aston.testproj.util.Constants.NEGATIVE_DEPOSIT;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
@@ -21,5 +23,5 @@ public class AccountDepositDto {
 
     @NotNull(message = EMPTY_DEPOSIT_VALUE)
     @Min(value = 0, message = NEGATIVE_DEPOSIT)
-    protected Long deposit;
+    protected BigDecimal deposit;
 }
